backend/requirements.txt
uuid~=1.30
fastapi~=0.111.0
pydantic~=2.8.2
uvicorn[standard]~=0.30.1
starlette~=0.37.2
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31
gunicorn~=22.0.0
psycopg2-binary~=2.9.9
